Move chart generation code into dedicated library
