+++ KustoLoco.sln
@@ -96,6 +96,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Intellisense", "libraries\I
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IntellisenseTests", "test\IntellisenseTests\IntellisenseTests.csproj", "{3D6E9D5C-2762-4FEB-AB4A-DF54F787E43F}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ScottPlotRendering", "libraries\ScottPlotRendering\ScottPlotRendering.csproj", "{5F3F34FE-3B2A-4272-A8FD-AF73E496AAD3}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -226,6 +228,10 @@ Global
 		{3D6E9D5C-2762-4FEB-AB4A-DF54F787E43F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3D6E9D5C-2762-4FEB-AB4A-DF54F787E43F}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3D6E9D5C-2762-4FEB-AB4A-DF54F787E43F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5F3F34FE-3B2A-4272-A8FD-AF73E496AAD3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5F3F34FE-3B2A-4272-A8FD-AF73E496AAD3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5F3F34FE-3B2A-4272-A8FD-AF73E496AAD3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5F3F34FE-3B2A-4272-A8FD-AF73E496AAD3}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -263,6 +269,7 @@ Global
 		{376ACB94-3327-4DE1-9F18-24C1EC896200} = {E515B24C-26F5-4DB4-893F-B581B23891E2}
 		{30D22373-FF6A-4669-BDB8-920718445C16} = {E515B24C-26F5-4DB4-893F-B581B23891E2}
 		{3D6E9D5C-2762-4FEB-AB4A-DF54F787E43F} = {932C804C-F6B7-4361-8360-E5238938E464}
+		{5F3F34FE-3B2A-4272-A8FD-AF73E496AAD3} = {E515B24C-26F5-4DB4-893F-B581B23891E2}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {2C5C9CC8-3EB2-4272-BBE5-637753D5E739}
