+++ applications/lokqlDx/lokqlDx.csproj
@@ -38,6 +38,7 @@
   <ItemGroup>
     <ProjectReference Include="..\..\libraries\Intellisense\Intellisense.csproj" />
     <ProjectReference Include="..\..\libraries\lokql-engine\lokql-engine.csproj" />
+    <ProjectReference Include="..\..\libraries\ScottPlotRendering\ScottPlotRendering.csproj" />
   </ItemGroup>
 
 </Project>
