+++ libraries/ScottPlotRendering/GenericTick.cs
@@ -1,0 +1,3 @@
+﻿namespace KustoLoco.ScottPlotRendering;
+
+public readonly record struct GenericTick(double Value, string Name);
