+++ libraries/ScottPlotRendering/IAxisLookup.cs
@@ -1,0 +1,8 @@
+﻿namespace KustoLoco.ScottPlotRendering;
+
+public interface IAxisLookup
+{
+    public double ValueFor(object? o);
+    string GetLabel(double position);
+    Dictionary<double, string> Dict();
+}
